[Issue-1988] Update UI Connect dApp screen
